core: move PathfindingResultConverter methods to appropriate classes
